Cow WIP
All tests pass, everything works. There is some minor cognitive overhead as a result of the lifetimes everywhere but it should also allow zero copy deserialization and streaming where possible, greatly reducing necessary allocations.

In most cases the code does not change. The one example that did, `strawberry.rs` we ended up reducing in size. A dedicated error type was not necessary for such an example.