Revert change to `request_custom`
I don't think there is a need to check for an empty response.

In the cases where the API returns an error response, the status code should not be `OK`, in which case we deserialize to an error.

In cases where the status code is OK and the response body is somehow empty, this is a bug in the API.