Chain Of Thought WIP
Some preliminary support for parsing chain of thought tags (`<thinking>` and so on). The plan is to integrate this into markdown and HTML formatting so the thinking can be placed in it's own elements and formatted nicely with CSS (or hidden entirely).

`Block` text is separated into `Thought` and `Speech` with zero-copy iterators over the text yielding these elements. A `Thinkable` trait (probably not the final name) is added and implemented for `Block`, `Content` and `prompt::Message` with a plan for `response::Message` as well.

This is the second or third draft of this feature and the first one I'm somewhat satisfied with. Attention is paid to cases where the Assistant might forget to close a tag (or open one) to avoid unintended thought leakage if desired.