CoT Markdown EOD WIP
Everything mostly works but there are some issues related to rendering of the markdown (missing newlines) and html is missing the `class`. Need to step though and see why `options.attrs` seems to not be working. It's supposed to be set to true when calling `Html` but it's not working.d