Memory Palace Refactor WIP
## Warning: This will not build (yet)

This is a major, major, reworking of the Palace. Claude 4 and I are leaning hard into the metaphor and using the language of a text adventure game to aid the memory retrieval agent.

The goal is to make something fun and intuitive while still being powerful. We're planning on leveraging a hybrid of graph and vector search and have created sets of tools for read and write agents.

For example, the retrieval agent, "Navigator" has a `Map` and can `Walk` in various directions. `Recall` allows global search. The agent's mind wanders. The `Archivist` will probably change quite a bit.

Next is to implement the rest of the navigator's methods, the Archivist, get it building, write unit tests, and integrate the tool into chat.